public/ocr-test/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
Pillow
streamlit
pandas
